src/requirements.txt
Flask==2.2.0
Markdown==3.4.3
Flask_WTF==1.1.1
Requests==2.31.0
Werkzeug==2.2.2
matplotlib==3.7.1
numpy==1.23.5
Pillow==9.5.0
tensorflow==2.12.0